corrected tests

+++ spec/factories/log.rb
@@ -8,7 +8,7 @@ FactoryGirl.define do
 
     message { "[#{Time.now.to_s}] [#{[200,404,500,422,400,201].sample}] #{Faker::Internet.url}" }
 
-    sequence(:logged_at) { |n| Time.now.to_f + (n * sequence_step) }
+    sequence(:logged_at) { |n| Time.now.to_i + (n * sequence_step) }
 
     initialize_with { attributes }
   end
